Updating Code Examples

+++ java-environment/day11/src/keywordthis/Account.java
@@ -37,7 +37,7 @@ public class Account {
 		count++;
 	}
 	
-	public void deposit(int amount) {
+	public void deposit(double amount) {
 		/*
 		 * Here, the keyword 'this' is not required
 		 * since there are no other variables that
